Updated CryptoEditForm and CryptoNewForm

+++ front-end/src/Components/CryptoNewForm.js
@@ -74,7 +74,7 @@ function CryptoNewForm() {
                 <label htmlFor="price">Price:</label>
                 <input 
                     id="price"
-                    type="integer"
+                    type="numeric"
                     name="price"
                     value={crypto.price}
                     placeholder="0"
@@ -83,7 +83,7 @@ function CryptoNewForm() {
                 <label htmlFor="rating">Rating:</label>
                 <input 
                     id="rating"
-                    type="integer"
+                    type="numeric"
                     name="rating"
                     value={crypto.rating}
                     placeholder="0"
